# Performance degradation for v1.6.15 fix