--- a/dataRetriever/factory/resolverscontainer/metaResolversContainerFactory.go
+++ b/dataRetriever/factory/resolverscontainer/metaResolversContainerFactory.go
@@ -27,7 +27,12 @@ func NewMetaResolversContainerFactory(
 		args.Marshalizer = marshal.NewSizeCheckUnmarshalizer(args.Marshalizer, args.SizeCheckDelta)
 	}
 
-	thr, err := throttler.NewNumGoRoutinesThrottler(args.NumConcurrentResolvingJobs)
+	mainThrottler, err := throttler.NewNumGoRoutinesThrottler(args.NumConcurrentResolvingJobs)
+	if err != nil {
+		return nil, err
+	}
+
+	trieNodesThrottler, err := throttler.NewNumGoRoutinesThrottler(args.NumConcurrentResolvingTrieNodesJobs)
 	if err != nil {
 		return nil, err
 	}
@@ -46,7 +51,8 @@ func NewMetaResolversContainerFactory(
 		triesContainer:                  args.TriesContainer,
 		inputAntifloodHandler:           args.InputAntifloodHandler,
 		outputAntifloodHandler:          args.OutputAntifloodHandler,
-		throttler:                       thr,
+		throttler:                       mainThrottler,
+		trieNodesThrottler:              trieNodesThrottler,
 		isFullHistoryNode:               args.IsFullHistoryNode,
 		mainPreferredPeersHolder:        args.MainPreferredPeersHolder,
 		fullArchivePreferredPeersHolder: args.FullArchivePreferredPeersHolder,