--- a/dataRetriever/factory/resolverscontainer/shardResolversContainerFactory_test.go
+++ b/dataRetriever/factory/resolverscontainer/shardResolversContainerFactory_test.go
@@ -100,8 +100,15 @@ func TestNewShardResolversContainerFactory_NewNumGoRoutinesThrottlerFailsShouldE
 
 	args := getArgumentsShard()
 	args.NumConcurrentResolvingJobs = 0
-	rcf, err := resolverscontainer.NewShardResolversContainerFactory(args)
 
+	rcf, err := resolverscontainer.NewShardResolversContainerFactory(args)
+	assert.Nil(t, rcf)
+	assert.Equal(t, core.ErrNotPositiveValue, err)
+
+	args.NumConcurrentResolvingJobs = 10
+	args.NumConcurrentResolvingTrieNodesJobs = 0
+
+	rcf, err = resolverscontainer.NewShardResolversContainerFactory(args)
 	assert.Nil(t, rcf)
 	assert.Equal(t, core.ErrNotPositiveValue, err)
 }
@@ -465,21 +472,22 @@ func TestShardResolversContainerFactory_IsInterfaceNil(t *testing.T) {
 
 func getArgumentsShard() resolverscontainer.FactoryArgs {
 	return resolverscontainer.FactoryArgs{
-		ShardCoordinator:                mock.NewOneShardCoordinatorMock(),
-		MainMessenger:                   createMessengerStubForShard("", ""),
-		FullArchiveMessenger:            createMessengerStubForShard("", ""),
-		Store:                           createStoreForShard(),
-		Marshalizer:                     &mock.MarshalizerMock{},
-		DataPools:                       createDataPoolsForShard(),
-		Uint64ByteSliceConverter:        &mock.Uint64ByteSliceConverterMock{},
-		DataPacker:                      &mock.DataPackerStub{},
-		TriesContainer:                  createTriesHolderForShard(),
-		SizeCheckDelta:                  0,
-		InputAntifloodHandler:           &mock.P2PAntifloodHandlerStub{},
-		OutputAntifloodHandler:          &mock.P2PAntifloodHandlerStub{},
-		NumConcurrentResolvingJobs:      10,
-		MainPreferredPeersHolder:        &p2pmocks.PeersHolderStub{},
-		FullArchivePreferredPeersHolder: &p2pmocks.PeersHolderStub{},
-		PayloadValidator:                &testscommon.PeerAuthenticationPayloadValidatorStub{},
+		ShardCoordinator:                    mock.NewOneShardCoordinatorMock(),
+		MainMessenger:                       createMessengerStubForShard("", ""),
+		FullArchiveMessenger:                createMessengerStubForShard("", ""),
+		Store:                               createStoreForShard(),
+		Marshalizer:                         &mock.MarshalizerMock{},
+		DataPools:                           createDataPoolsForShard(),
+		Uint64ByteSliceConverter:            &mock.Uint64ByteSliceConverterMock{},
+		DataPacker:                          &mock.DataPackerStub{},
+		TriesContainer:                      createTriesHolderForShard(),
+		SizeCheckDelta:                      0,
+		InputAntifloodHandler:               &mock.P2PAntifloodHandlerStub{},
+		OutputAntifloodHandler:              &mock.P2PAntifloodHandlerStub{},
+		NumConcurrentResolvingJobs:          10,
+		NumConcurrentResolvingTrieNodesJobs: 3,
+		MainPreferredPeersHolder:            &p2pmocks.PeersHolderStub{},
+		FullArchivePreferredPeersHolder:     &p2pmocks.PeersHolderStub{},
+		PayloadValidator:                    &testscommon.PeerAuthenticationPayloadValidatorStub{},
 	}
 }