## Reasoning behind the pull request
- the number of go routines that the node launched for resolving the trie nodes requests interfered with the processing go routine
  
## Proposed changes
- added a critical section in the trie nodes resolvers and a different throttler
- the results for trie sync performance benchmark was carried on a small & decoupled network containing mainnet data containing a full squad and a node from shard 0 that performed a trie sync. This is the worse case scenario as there is a single node that responds to trie requests data and the output antiflooder will severely limit the trie node responses.


The trie sync result when running with v1.6.15 version is this:
```
INFO [2024-01-31 16:28:36.340]   finished trie sync                       name = user accounts for shard 0 time elapsed = 40m59s num processed = 15668259 num large nodes = 0 num missing = 0 state data size = 2.16 GB total iterations = 4904541 total CPU time = 1h16m58.021939181s peak processing speed = 3.53 MB/s average processing speed = 919.48 KB/s 
```


The trie sync result when running with this branch is the following:
```
INFO [2024-01-31 17:18:46.782]   finished trie sync                       name = user accounts for shard 0 time elapsed = 44m30s num processed = 15668259 num large nodes = 0 num missing = 0 state data size = 2.16 GB total iterations = 5194380 total CPU time = 1h14m28.877860401s peak processing speed = 2.61 MB/s average processing speed = 846.79 KB/s 
```

## Testing procedure
- standard system test

## Pre-requisites

Based on the [Contributing Guidelines](https://github.com/multiversx/mx-chain-go/blob/master/.github/CONTRIBUTING.md#branches-management) the PR author and the reviewers must check the following requirements are met:
- was the PR targeted to the correct branch?
- if this is a larger feature that probably needs more than one PR, is there a `feat` branch created?
- if this is a `feat` branch merging, do all satellite projects have a proper tag inside `go.mod`?